// src/generated/models/Cfm_manualentryauditsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cfm_manualentryauditsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cfm_manualentryauditsstatecode = keyof typeof Cfm_manualentryauditsstatecode;
export const Cfm_manualentryauditsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cfm_manualentryauditsstatuscode = keyof typeof Cfm_manualentryauditsstatuscode;

export interface Cfm_manualentryauditsBase {
  "cfm_ARAging@odata.bind"?: string;
  cfm_fieldname: string;
  cfm_manualentryauditid: string;
  cfm_newvalue?: number;
  cfm_oldvalue?: number;
  cfm_planperiod?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Cfm_manualentryauditsstatecode;
  statuscode?: Cfm_manualentryauditsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cfm_manualentryaudits extends Cfm_manualentryauditsBase {
  cfm_aragingname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  cfm_araging?: object;
  _cfm_araging_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}